+++ examples/full_example/lib/widgetbook.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:widgetbook/next.dart' as next;
+import 'package:widgetbook/next.dart';
 import 'package:widgetbook/widgetbook.dart';
 
 import 'components.book.dart';
@@ -18,12 +18,12 @@ class WidgetbookApp extends StatelessWidget {
     return Widgetbook.material(
       directories: components,
       addons: [
-        DeviceFrameAddon(devices: Devices.ios.all),
+        DeviceFrameAddon(Devices.ios.all),
         InspectorAddon(),
         GridAddon(100),
         AlignAddon(),
         ZoomAddon(),
-        next.ThemeAddon<AppThemeData>(
+        ThemeAddon<AppThemeData>(
           {
             'Blue': AppThemeData(Colors.blue),
             'Yellow': AppThemeData(Colors.yellow),