+++ packages/widgetbook/lib/src/addons/accessibility_addon/addon.dart
@@ -1,1 +1,0 @@
-export 'accessibility_addon.dart';