+++ packages/widgetbook/lib/src/addons/alignment_addon/addon.dart
@@ -1,1 +1,0 @@
-export 'alignment_addon.dart';