+++ packages/widgetbook/lib/src/addons/device_frame_addon/addon.dart
@@ -1,5 +1,0 @@
-export 'package:device_frame/device_frame.dart';
-
-export 'device_frame_addon.dart';
-export 'device_frame_setting.dart';
-export 'none_device.dart';