+++ packages/widgetbook/lib/src/addons/inspector_addon/addon.dart
@@ -1,1 +1,0 @@
-export 'inspector_addon.dart';