+++ packages/widgetbook/lib/src/addons/localization_addon/addon.dart
@@ -1,1 +1,0 @@
-export 'localization_addon.dart';