+++ packages/widgetbook/lib/src/addons/text_scale_addon/addon.dart
@@ -1,1 +1,0 @@
-export 'text_scale_addon.dart';