+++ packages/widgetbook/lib/src/addons/zoom_addon/addon.dart
@@ -1,1 +1,0 @@
-export 'zoom_addon.dart';