+++ packages/widgetbook/lib/src/next/addons/text_scale_addon.dart
@@ -6,9 +6,12 @@
 import 'package:flutter/widgets.dart';
 
 import '../../fields/fields.dart';
+import 'base/addon.dart';
 import 'base/mode.dart';
 import 'base/mode_addon.dart';
 
+/// An [Addon] for changing the active [MediaQueryData.textScaleFactor]
+/// via [MediaQuery].
 class TextScaleMode extends Mode<double> {
   TextScaleMode(super.value);