+++ packages/widgetbook/lib/src/next/addons/zoom_addon.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/widgets.dart';
 
 import '../../fields/fields.dart';
+import 'base/addon.dart';
 import 'base/mode.dart';
 import 'base/mode_addon.dart';
 
@@ -16,6 +17,7 @@ class ZoomMode extends Mode<double> {
   }
 }
 
+/// An [Addon] for zoom/scaling.
 class ZoomAddon extends ModeAddon<double> {
   ZoomAddon()
       : super(