+++ packages/widgetbook/test/src/integrations/widgetbook_cloud_integration_test.dart
@@ -1,6 +1,6 @@
 import 'package:flutter_test/flutter_test.dart';
 import 'package:mocktail/mocktail.dart';
-import 'package:widgetbook/next.dart' hide AlignmentAddon;
+import 'package:widgetbook/next.dart';
 import 'package:widgetbook/widgetbook.dart';
 
 import '../../helper/helper.dart';